[#120] refactor: Oauth 요청 로직 리팩토링
