+++ src/components/auth/OAuth.tsx
@@ -2,43 +2,19 @@
 
 import { useRouter, useSearchParams } from "next/navigation";
 
-import { OAuthApiRequest } from "@/types/auth";
-import { useUserActions } from "@/store/user";
-import { oAuthApi, signInApi } from "@/api/modules/user";
 import { useRunOnce } from "@/hooks/common";
-import useAuthSession from "@/hooks/useAuthStorage";
+import { useAuth } from "@/hooks/auth";
 
 export default function OAuth() {
   const router = useRouter();
-  const code = useSearchParams().get("code") || "";
+  const { signIn } = useAuth();
 
-  const { setUser } = useUserActions();
-  const authSession = useAuthSession();
+  const code = useSearchParams().get("code") || "";
 
   useRunOnce(() => {
     const fetch = async () => {
-      const params: OAuthApiRequest = {
-        providerType: "kakao",
-        redirectUri: window.location.origin + window.location.pathname,
-        code,
-      } as const;
       try {
-        const { registered, memberInfo } = await oAuthApi(params);
-
-        if (registered) {
-          const data = await signInApi({
-            providerType: memberInfo.providerType!,
-            providerToken: memberInfo.providerToken,
-          });
-
-          setUser(data);
-          authSession.set(data);
-
-          router.push("/");
-          return;
-        }
-        setUser(memberInfo);
-        router.push("/signUp");
+        signIn(code);
       } catch (err) {
         if (err instanceof Error) {
           if (err.message) {
